consolidate response_test code
